+++ nafas/params.py
@@ -94,7 +94,7 @@ STANDARD_MENU = {
         14: "Balancing",
         15: "Energizing",
         16: "Box",
-        17: "Coherent Breathing"},
+        17: "Coherent"},
     "level": {
         1: "Beginner",
         2: "Medium",
@@ -421,7 +421,7 @@ BOX_ADVANCED = {
     "cycle": 15}
 
 
-COHERENT_BREATHING_BEGINNER = {
+COHERENT_BEGINNER = {
     "ratio": [
         5,
         0,
@@ -432,7 +432,7 @@ COHERENT_BREATHING_BEGINNER = {
     "cycle": 30}
 
 
-COHERENT_BREATHING_MEDIUM = {
+COHERENT_MEDIUM = {
     "ratio": [
         5,
         0,
@@ -443,7 +443,7 @@ COHERENT_BREATHING_MEDIUM = {
     "cycle": 50}
 
 
-COHERENT_BREATHING_ADVANCED = {
+COHERENT_ADVANCED = {
     "ratio": [
         5,
         0,
@@ -502,9 +502,9 @@ PROGRAMS = {"Clear Mind": {"Beginner": CLEAR_MIND_BEGINNER,
             "Box": {"Beginner": BOX_BEGINNER,
                     "Medium": BOX_MEDIUM,
                     "Advanced": BOX_ADVANCED},
-            "Coherent Breathing": {"Beginner": COHERENT_BREATHING_BEGINNER,
-                                  "Medium": COHERENT_BREATHING_MEDIUM,
-                                  "Advanced": COHERENT_BREATHING_ADVANCED},
+            "Coherent": {"Beginner": COHERENT_BEGINNER,
+                                  "Medium": COHERENT_MEDIUM,
+                                  "Advanced": COHERENT_ADVANCED},
         }