+++ test/test.py
@@ -113,7 +113,7 @@ True
 14- Balancing (~ 2 minutes)
 15- Energizing (~ 2 minutes)
 16- Box (~ 2 minutes)
-17- Coherent Breathing (~ 8 minutes)
+17- Coherent (~ 8 minutes)
 - Choose a level:
 <BLANKLINE>
 1- Beginner